chore: rename test case and update examples from class component to function
